+++ ui/src/tugboat/error/index.ts
@@ -39,3 +39,15 @@ export class UnknownStateError extends Error {
         super(`Unknown state: ${state.constructor.name}`);
     }
 }
+
+export class UngroupedExamplesError extends Error {
+    constructor() {
+        super(`The examples aren't grouped but are expected to be.`);
+    }
+}
+
+export class GroupedExamplesError extends Error {
+    constructor() {
+        super(`The examples are grouped but are expected not to be.`);
+    }
+}
