+++ ui/src/tugboat/lib/Example.ts
@@ -14,15 +14,3 @@ export function isGroupedExamples(
     }
     return true;
 }
-
-export class UngroupedExamplesError extends Error {
-    constructor() {
-        super(`The examples aren't grouped but are expected to be.`);
-    }
-}
-
-export class GroupedExamplesError extends Error {
-    constructor() {
-        super(`The examples are grouped but are expected not to be.`);
-    }
-}
